Fixed partition creation on first boot after flash nuke

+++ ports/rp2/modules/_boot_fat.py
@@ -4,11 +4,18 @@ import machine, rp2
 
 # Try to mount the filesystem, and format the flash if it doesn't exist.
 bdev = rp2.Flash()
+partName = bytearray(16)
+bdev.readblocks(0, partName)
+partitionExists = partName[3:11] == "MSDOS5.0"
+
 try:
-    fs = vfs.VfsFat(bdev)
+    if partitionExists:
+        fs = vfs.VfsFat(bdev)
+    else:
+        vfs.VfsFat.mkfs(bdev)
+        fs = vfs.VfsFat(bdev)
 except:
-    vfs.VfsFat.mkfs(bdev)
-    fs = vfs.VfsFat(bdev)
+    pass
 vfs.mount(fs, "/")
 
-del vfs, bdev, fs
+del vfs, bdev, fs, partName, partitionExists
